Add function to append colored class to gridBoxes
Added an event listener to each div.grid.box element, that will call
the addColorClass function whenever the user's mouse pointer hovers
over one of the elements. The addColorClass function adds the
'colored' class to the elements, which will be used in the stylesheet
to change the background color of the elements.