Create function to change resolution of grid
Created function, changeResolution, with no parameters, that allows the
user to enter a new value for the rowPx and colPx values of the grid,
then generated the grid again with new values, and reassigns the
function to draw on each square.

Also changed the rowPx and colPx constants to variables so that it can
be updated.